Added filtering capabilities based on two main categories
1. TFA Methods -- sms, email, phone, hw, sw.
2. No TFA -- all non-tfa sites, non-tfa but in-progress

These filters play nicely with search. In fact, some of
their functionality relies on the help of the search plugin.
Currently everything works well EXCEPT for the checkboxes
being disabled. I'm not sure if there is an issue with this
version of semantic-ui or not, but disabling the checkboxes
does not seem to be preventing them from being clicked on as
is suggested in the documentation. They also don't respond
to the same events as described in the documentation. I suggest
The site be updated to use the current version of semantic-ui
and then these changes be tested and considered against that.