Update Paypal doc
The old Paypal doc did not require sign-in; it described the Paypal SMS key and linked to the key setup page, but did not describe or the software (Symantec VIP) or hardware keys or provide any way to set up software or hardware keys.  This new link requires sign-in but provides buttons to set up SMS, software, or hardware keys.