Add Regions Bank (#1875)
* Added Regions Bank

Added Regions Bank to list of banks.  They do not currently support two factor authentication.

* Added Regions Bank Logo

Added RegionsBank.png to be used as a logo

* Added Logo to Regions Bank Listing

Added a link to the RegionsBank.png image.

* Replaced Regions Bank Logo

Replaced Regions Bank Logo with a 32x32 png. The old one was only 16x16.

* Delete RegionsBank.png

This was accidentally added to the wrong directory.  I'm deleting it.

* Changed RegionsBank image file extention

changed from *.img to *.png.

* Changed URL to HTTPS

Changed URL to HTTPS for security.