Add Minds
I was unable to find a suitable doc link.

"Minds (Minds, Inc.) is a free and open source social networking
platform that rewards you for your activity online with revenue, digital
currency and more views on your content. We are built on a foundation of
freedom, privacy and democratization." - About,
https://www.minds.com/p/about