small tweak to allow us to get at least the JS compression on our beta site.
We could push our _site folder to github pages branch but that means overwriting the .gitignore for our deploy to there. Instead, we can at least pull back the minified copies of files that are not generated on the fly, so that we at least get those benefits.

In the future I may make the additions into a bash script or rake option so that we can have it just copy back to root any file that doesnt end in .min.js thats in the JS folder