This is the next set of changes, brings in a lot, but also not yet complete. Its going to add some interesting things, though.
First its the real start of using the command line to make additions, which is not yet complete as I am working on the traversal of the schema to dictate the questions so its always up to date.

It moves the _sass folder inside the includes folder. Why? Well its simple, so we can include critical CSS in the future without moving the rest of our SCSS and SASS apart from each other. Its odd that they make it so difficult but this greatly enables us to speed up page loads in the future be cherry-picking out critical CSS.