method refactor of rakefile
I noticed in my own repo that it makes a bit more sense to call HTMLProofer from a single place, and the overlapping settings be configured as defaults. This makes it a bit easier to manage if you add to your configuration, but also to understand what the expected difference between tests is without having to mentally keep track of what is common between them.

I threw in a small change to move the require for the HTML Proofer to the method itself. By doing this, it gets loaded only if that method gets called, which means rake tasks will not fail for windows machines that commonly do not have libcurl properly installed, making your rake tasks fail because it cannot load up the HTML Proofer library properly even though it installed the ruby library, because the ruby library has a dependency on something annoying and likely not configured for windows based machines. The change however does not any other effect on its ability to run, so its a rather safe change