Reverts the value of all doc fields to URLs
Some of the service entries contain a relative path in
their doc field that points to a local page on
twofactorauth.org. This paradigm is followed for sites
that support 2FA, but do not have good documentation articles
on their websites.

The problem that this commit addresses is for external tools
and services that rely on the doc field to present the user
with a URL to access the 2FA documentation for a service.

For example:
- 1Password
- 2fanotifier.org

These tools will have to specifically address the scenarios
where the doc field is a relative path instead of a full URL.
Updating the source data allows third parties to use the data
without addressing edge cases.