Added exceptions for QNAP and Synology
We got some reports at 1Password regarding QNAP and Synology being listed as 2FA enabled when a user creates a login for a site like QNAP forums (a subdomain at https://forum.qnap.com/), but the other websites do not allow for it. Turns out, this is because the 2FA is associated with a local IP address or custom domain the NAS is hosted on.