Update gmail.com to mail.google.com (#3247)
There are many URLs in the data.json data set that
redirect before landing on the final site. This
causes problems for third party clients (e.g. 2FA Notifier)
because they match on the origin URL.

Updating the data to reflect the final URL allows these
third party services to function correctly while also
preserving the functionality of twofactorauth.org
for end-users of the website.

We need to be careful when working with temporary redirects
since they could change in the future, but this PR is
just for gmail, which has a very VERY low likelihood of
redirecting to a new subdomain any time soon.