Add Mint doc (#3507)
* Update Mint entry

MFA for Mint must be configured via another Intuit site, since Mint itself doesn't support setting it up. I added a link to the TurboTax documentation which describes how to do this. Unfortunately this is rather confusing so it would be nice to add an extra note describing the process, but the current schema doesn't seem to allow that.

* Update doc link for Mint based on PR suggestion