Ampersand correction
When a site name contains and ampersand (&) the normal HTML-notation would be to use &amp;. Yet, this &amp; will show up in the JSON-export and not the &.
Other examples in the data show, that the ampersand can be used without the HTML notation in the name when it's in single quotation marks (').
All names with ampersands are now in single quotes.