Second version of attention messages
Made the Attention-message dismissable. Desktop table keeps the Attention-message as an exception popup over the buttons, where the mobile table has a div between the icon and the entries table that's also dismissable. Changed the wording of the buttons on the mobile table ("Ask for it on ...") and subsequently moved email-button in a new line because the looked crammed when all three contact methods where available. The link on the desktop table now comes from the desktop_publicity.html include to keep the desktop-table.html less messy.