Update dashlane.com.json
Dashlane does not currently support U2F for public use following the transition to using the webapp and removing their desktop app (the desktop app can be installed, but then refuses to run once the user logs in and instead directs them to the webapp). As such, it is no longer possible (despite advertising saying otherwise) to use U2F with dashlane.