2fa that's compromised
Microsoft s 2fa has been acting very strange to me. password s are wrong it keeps asking to make new ones and pop ups when I search for things asking for password s when I'm all ready logged in. Microsoft asked me to change password on 3 account s when  I was logged in all ready on there doc page it said the app hadn't been updated since 2015 and was being discontinued these all together seams like issues that with the alert today if I had a dollar I would bet it on the 2fa from Microsoft has become a backdoor access chanel but not really sure how to go about testing this. and looking at my logs there are a couple ip address that are not in the location that they should be and with different ip s this could just be a VPN thing but but like 1 out of 5 randomly different seams out of place to me so I thought I should let someone know.